+++ core/src/test/java/cucumber/runtime/formatter/JUnitFormatterTest.java
@@ -12,6 +12,7 @@ import cucumber.runtime.snippets.FunctionNameGenerator;
 import gherkin.pickles.PickleStep;
 import org.custommonkey.xmlunit.Diff;
 import org.custommonkey.xmlunit.XMLUnit;
+import org.junit.AssumptionViolatedException;
 import org.junit.Test;
 import org.xml.sax.SAXException;
 
@@ -22,6 +23,8 @@ import java.io.FileInputStream;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.util.AbstractMap.SimpleEntry;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -89,6 +92,40 @@ public class JUnitFormatterTest {
                 "When second step............................................................passed\n" +
                 "Then third step.............................................................passed\n" +
                 "]]></system-out>\n" +
+                "    </testcase>\n" +
+                "</testsuite>\n";
+        assertXmlEqual(expected, formatterOutput);
+    }
+
+    @Test
+    public void should_format_skipped_scenario() throws Throwable {
+        CucumberFeature feature = TestHelper.feature("path/test.feature",
+                "Feature: feature name\n" +
+                        "  Scenario: scenario name\n" +
+                        "    Given first step\n" +
+                        "    When second step\n" +
+                        "    Then third step\n");
+        Map<String, Result> stepsToResult = new HashMap<String, Result>();
+        Throwable exception = new AssumptionViolatedException("message");
+        stepsToResult.put("first step", result("skipped", exception));
+        stepsToResult.put("second step", result("skipped"));
+        stepsToResult.put("third step", result("skipped"));
+        long stepDuration = milliSeconds(1);
+
+        String formatterOutput = runFeatureWithJUnitFormatter(feature, stepsToResult, stepDuration);
+
+        String stackTrace = getStackTrace(exception);
+        String expected = "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?>\n" +
+                "<testsuite failures=\"0\" name=\"cucumber.runtime.formatter.JUnitFormatter\" skipped=\"1\" tests=\"1\" time=\"0.001\">\n" +
+                "    <testcase classname=\"path/test.feature\" name=\"scenario name\" time=\"0.001\">\n" +
+                "        <skipped message=\"" + stackTrace.replace("\n\t", "&#10;&#9;") + "\"><![CDATA[" +
+                "Given first step............................................................skipped\n" +
+                "When second step............................................................skipped\n" +
+                "Then third step.............................................................skipped\n" +
+                "\n" +
+                "StackTrace:\n" +
+                stackTrace +
+                "]]></skipped>\n" +
                 "    </testcase>\n" +
                 "</testsuite>\n";
         assertXmlEqual(expected, formatterOutput);
@@ -542,5 +579,11 @@ public class JUnitFormatterTest {
 
     private Long milliSeconds(int milliSeconds) {
         return milliSeconds * 1000000L;
+    }
+
+    private String getStackTrace(Throwable exception) {
+        StringWriter sw = new StringWriter();
+        exception.printStackTrace(new PrintWriter(sw));
+        return sw.toString();
     }
 }